# channel: remove parking_lot dependency