--- a/crossbeam-channel/Cargo.toml
+++ b/crossbeam-channel/Cargo.toml
@@ -16,7 +16,6 @@ keywords = ["channel", "mpmc", "select", "golang", "message"]
 categories = ["algorithms", "concurrency", "data-structures"]
 
 [dependencies]
-parking_lot = "0.7"
 smallvec = "0.6.2"
 
 [dependencies.crossbeam-utils]