--- a/crossbeam-channel/src/lib.rs
+++ b/crossbeam-channel/src/lib.rs
@@ -348,7 +348,6 @@
 #![warn(missing_debug_implementations)]
 
 extern crate crossbeam_utils;
-extern crate parking_lot;
 extern crate smallvec;
 
 mod channel;