@stjepang 2019-01-18
@Amanieu Is there perhaps a reasonably simple alternative approach we could use here?

But anyhow, I'm not worried about these spinlocks, for several reasons:

1. Channels already have blocking sections of code in multiple places, where we occasionally (on cold pahts) have to spin in a similar way. It doesn't appear to be a problem in practice.

2. Mutexes are only used when the channel is full or empty and the operation is going to block with `thread::park()` anyway. It's not really a contentious situation.

3. The only scenario where threads will put really high pressure on mutexes is when channel capacity is small (e.g. 0 or 1) and threads are sending a ton of messages through the channel. But this is an antipattern anyway because such channels are supposed to be used for lightweight thread synchronization rather than sending lots of messages.